Update Metal Example to work correctly
This example failed to load metal due to it missing the SetGPULayers
option.

I also updated the instructions to use go build instead of go run. This
is due to go run not loading the ggml-metal.metal file. I was unable to
get this working. Compiling it and running ./main will load the
ggml-metal.metal file in the CWD.