Fix Stopwords/Antiprompt
The original break was breaking out of the inner for loop, but not the outer while loop. Repurpose (?) the existing unused variable is_antiprompt to terminate the while loop if any antiprompt was found. We can still break out of the for loop, since we already found one antiprompt, finding others is not useful.

Signed-off-by: 65a <65a@63bit.net>